Move non-junit4 test API implementation into ui-test
