+++ compose/ui/ui-test/src/desktopMain/kotlin/androidx/compose/ui/test/DesktopSynchronization.desktop.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 The Android Open Source Project
+ * Copyright 2023 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package androidx.compose.ui.test.junit4
+package androidx.compose.ui.test
 
 import java.util.concurrent.ExecutionException
 import java.util.concurrent.FutureTask
