+++ compose/ui/ui-test/src/nativeMain/kotlin/androidx/compose/ui/test/Synchronization.nativeMain.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 The Android Open Source Project
+ * Copyright 2023 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,9 +14,8 @@
  * limitations under the License.
  */
 
-package androidx.compose.ui.test.junit4
+package androidx.compose.ui.test
 
-import androidx.compose.ui.test.NanoSecondsPerMilliSecond
 import kotlin.coroutines.suspendCoroutine
 import kotlinx.cinterop.cValue
 import kotlinx.coroutines.runBlocking
