+++ compose/ui/ui-test/src/skikoMain/kotlin/androidx/compose/ui/test/ComposeRootRegistry.skiko.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 The Android Open Source Project
+ * Copyright 2023 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package androidx.compose.ui.test.junit4
+package androidx.compose.ui.test
 
 import androidx.compose.ui.InternalComposeUiApi
 import androidx.compose.ui.platform.PlatformContext
@@ -31,7 +31,7 @@ internal class ComposeRootRegistry : PlatformContext.RootForTestListener {
     private val roots = mutableSetOf<PlatformRootForTest>()
 
     /**
-     * Returns if the registry is setup to receive registrations from [PlatformRootForTest]s
+     * Returns if the registry is set up to receive registrations from [PlatformRootForTest]s
      */
     private var isTracking = false
