+++ compose/ui/ui-test/src/skikoMain/kotlin/androidx/compose/ui/test/MainTestClockImpl.skikoMain.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 The Android Open Source Project
+ * Copyright 2020 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,6 +14,19 @@
  * limitations under the License.
  */
 
-package androidx.compose.ui.test.junit4
+package androidx.compose.ui.test
 
-internal actual inline fun <T> synchronized(lock: Any, block: () -> T) = block()
+import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.test.TestCoroutineScheduler
+
+@OptIn(ExperimentalCoroutinesApi::class)
+internal class MainTestClockImpl(
+    testScheduler: TestCoroutineScheduler,
+    frameDelayMillis: Long,
+    onTimeAdvanced: (Long) -> Unit
+) : AbstractMainTestClock(
+    testScheduler = testScheduler,
+    frameDelayMillis = frameDelayMillis,
+    runOnUiThread = ::runOnUiThread,
+    onTimeAdvanced = onTimeAdvanced
+)
